+++ mathesar_ui/src/stores/table-data/record-summaries/__tests__/recordSummary.test.ts
@@ -2,7 +2,7 @@ import { ImmutableMap } from '@mathesar-component-library';
 import {
   renderSummaryFromFieldsAndFkData,
   renderSummaryFromValuesObject,
-} from '../recordSummary';
+} from '../recordSummaryUtils';
 import type { DataForRecordSummaryInFkCell } from '../recordSummaryTypes';
 
 test.each(
